feat: refactor CSS directive handling by extracting utility functions for Tailwind reference detection and improving test case timeout
